api/news/requirements.txt
Flask
newsapi-python
flask-cors